🐛 Fix input's nested look ahead
Cause: the previous implementation didn't take into consideration that nested looking ahead was a possibility. Therefore, it would stop looking ahead altogether when stopping a nested lookahead.
Fix: model the nested lookahead as a stack of `LookAheadFrame`, which keeps track at where in the lookahead buffer the call started, and how many elements it has looked ahead. That way, whenever leaving a lookahead, we can unwind just the right amount.

+++ src/tests/number/count.rs
@@ -46,7 +46,7 @@ fn parse_count_one_different() {
 	let output = parser(&mut input).unwrap();
 	assert!(output.iter().all(|&x| x == 1)); // All values match the parser's.
 	assert_eq!(input.next_token(), Some(42)); // Input was consumed as much as possible.
-	assert_eq!(input.consumed_count(), (repeat_count + 1) as u32); // Input was left intact.
+	assert_eq!(input.consumed_count(), repeat_count + 1); // Input was left intact.
 }
 
 #[test]
